feat(@typed-storage/react): allow using type-safe localStorage outside the React context
